# src/docs/openapi.yaml
openapi: "3.0.0"
info:
  version: 6.0.0
  title: Jeera Project Manager
  license:
    name: MIT
  contact:
    name: Antoine Cheron
    url: https://antoinecheron.github.io
    email: antoine.cheron@irisa.fr
servers:
  - url: http://localhost:3000/
paths:
  /users:
    post:
      summary: Creates a new user (admin only)
      operationId: createUser
      x-@id: schema:RegisterAction
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User just created
          headers:
            Location:
              description: 'URL to the resource created'
              schema:
                x-@type: '@id'
                type: string
                format: url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
          links:
            confirmEmail:
              x-@relation: 
                - lcy:next
                - vnd_jeera:rel/confirmEmail
              operationId: confirmEmail
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/confirm: # to remove in v2
    get:
      summary: Confirm email ownership
      operationId: confirmEmail
      x-@id: vnd_jeera:confirmEmail
      tags: 
        - users
      parameters:
        - name: token
          x-@id: vnd_jeera:emailConfirmationToken
          in: path
          description: Confirmation token received by email
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Success
          links:
            login:
              operationId: login
              x-@relation: vnd_jeera:rel/login
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/login:
    post:
      summary: Login into the application
      operationId: login
      x-@id: vnd_jeera:login
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful connection
          headers:
            Authorization:
              description: Token to use
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginConfirmation'
          links:
            logout:
              operationId: logout
              x-@relation: vnd_jeera:rel/logout
            getUserDetails:
              operationId: getUserDetails
              x-@relation: vnd_jeera:rel/detail
            listProjects:
              operationId: listProjects
              x-@relation: vnd_jeera:rel/listProjects
        '403':
          description: 'Wrong credentials'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/logout:
    post:
      summary: Logout
      operationId: logout
      x-@id: vnd_jeera:logout
      tags: 
        - users
      responses:
        '204':
          description: 'Successfully logged out - revoke token'
        '500':
          description: 'Unexpected internal server error, try again'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /user:
    get:
      summary: Get a user details
      operationId: getCurrentUserDetails
      x-@id: vnd_jeera:getCurrentUserDetails
      tags:
        - users
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
          links:
            listProjects:
              operationId: listProjects
              x-@relation: vnd_jeera:rel/listProjects
            changePassword:
              operationId: updateUserPassword
              x-@relation: vnd_jeera:rel/changePassword
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /user/{userId}:
    parameters: 
      - $ref: '#/components/parameters/userId'
    get:
      summary: Get a user details
      operationId: getUserDetails
      x-@id: vnd_jeera:getUserDetails
      tags:
        - users
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /user/password:
    parameters: 
      - $ref: '#/components/parameters/userId'
    put:
      summary: Updates a user's password
      operationId: updateUserPassword
      x-@id: vnd_jeera:updateUserPassword
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '204':
          description: 'Password successfully updated'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /projects:
    get:
      summary: List the projects that the connected user can see
      operationId: listProjects
      x-@id: vnd_jeera:listProjects
      tags:
        - projects
      parameters:
        - name: offset
          x-@id: hydra:offset
          in: query
          description: The number of items to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            format: int8
            minimum: 0
            default: 0
        - name: limit
          x-@id: hydra:limit
          in: query
          description: Number of elements to get
          required: false
          schema:
            type: integer
            format: int8
            minimum: 1
            maximum: 40
            default: 3
        - name: public
          in: query
          x-@id: vnd_jeera:public
          description: Should the listed projects be public - listing private projects requires to be logged in
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsList'
          links:
            createProject:
              operationId: createProject
              x-@relation: vnd_jeera:rel/create
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Creates a project
      operationId: createProject
      x-@id: vnd_jeera:createProject
      tags:
        - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          $ref: '#/components/responses/projectDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    projectDetails:
      description: Project details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Project'
      links:
        inviteUser:
          operationId: inviteUsersToProject
          x-@relation: vnd_jeera:rel/inviteToCollaborate
          parameters:
            projectId: '$response.body#/projectId'
        createTechnicalStory:
          operationId: createTechnicalStory
          x-@relation: vnd_jeera:rel/create
          parameters:
            projectId: '$response.body#/id'
        createUserStory:
          operationId: createUserStory
          x-@relation: vnd_jeera:rel/create
          parameters:
            projectId: '$response.body#/id'
        listTasks:
          operationId: listProjectTasks
          x-@relation: vnd_jeera:rel/listProjectTasks
          parameters:
            projectId: '$response.body#/id'
        archive:
          operationId: archiveProject
          x-@relation: vnd_jeera:rel/archive
          parameters:
            projectId: '$response.body#/id'
        unarchive:
          operationId: unarchiveProject
          x-@relation: vnd_jeera:rel/unarchive
          parameters:
            projectId: '$response.body#/id'
        delete:
          operationId: deleteProject
          x-@relation: vnd_jeera:rel/delete
          parameters:
            projectId: '$response.body#/id'
        star:
          operationId: starProject
          x-@relation: vnd_jeera:rel/starProject
          parameters:
            projectId: '$response.body#/id'
        analytics:
          operationId: getAnalytics
          x-@relation: vnd_jeera:rel/analytics
          parameters:
            resourceId: '$response.body#/id'
  parameters:
    userId: 
      name: userId
      x-@id: vnd_jeera:userId
      in: path
      description: Id of the user to retrieve details for
      required: true
      schema:
        type: string
        format: url
        x-@type: '@id'
...
  schemas:
    CreateUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
    UserDetails:
      type: object
      properties:
        username:
          type: string
        id:
          type: string
          format: uuid
          x-@id: vnd_jeera:userId
          x-@type: 'mo:uuid'
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        starredProjects:
          type: array
          items:
            type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required: [ 'username', 'password' ]
    LoginConfirmation:
      type: object
      properties:
        token:
          type: string
          x-@id: vnd_jeera:JWT
    UpdatePasswordRequest:
      type: object
      properties:
        previousPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
      required: [ previousPassword, newPassword ]
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        isPublic:
          type: boolean
          default: false
      required: [ name ]
    ProjectsList:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    Project:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
          format: uuid
          x-@id: vnd_jeera:projectId
          x-@type: 'mo:uuid'
        isPublic:
          type: boolean
        isArchived:
          type: boolean
        collaborators:
          type: array
          items:
            type: string
            format: uuid
      required: [ name, id, isPublic, isArchived, collaborators ]
...
    UserRole:
      type: string
      enum: 
        - Product Owner
        - Developer
...
    Error:
      type: object
      properties:
        errorCode:
          type: string
          x-@id: schema:identifier
...